@@ -34,6 +34,6 @@ setup(
         "nbconvert>=5.5",
         "nbformat",
     ],
-    python_requires=">= 3.5",
+    python_requires=">= 3.6",
     package_data={"jupyter_sphinx": ["thebelab/*", "css/*"]},
 )
