@@ -4,6 +4,7 @@ from itertools import groupby, count
 from sphinx.errors import ExtensionError
 import nbformat
 from jupyter_client.kernelspec import get_kernel_spec, NoSuchKernel
+from pathlib import Path
 
 
 def blank_nb(kernel_name):
@@ -74,12 +75,14 @@ def sphinx_abs_dir(env, *paths):
     # output_directory / jupyter_execute / path relative to source directory
     # Sphinx expects download links relative to source file or relative to
     # source dir and prepended with '/'. We use the latter option.
-    return "/" + os.path.relpath(
-        os.path.abspath(
-            os.path.join(output_directory(env), os.path.dirname(env.docname), *paths)
-        ),
-        os.path.abspath(env.app.srcdir),
-    )
+    out_path = (output_directory(env) /  Path(env.docname).parent / Path(*paths)).resolve()
+      
+    if os.name == "nt":
+        # Can't get relative path between drives on Windows
+        return out_path.as_posix()
+
+    # Path().relative_to() doesn't work when not a direct subpath
+    return "/" + os.path.relpath(out_path, env.app.srcdir)
 
 
 def output_directory(env):
@@ -90,6 +93,4 @@ def output_directory(env):
 
     # Note: we are using an implicit fact that sphinx output directories are
     # direct subfolders of the build directory.
-    return os.path.abspath(
-        os.path.join(env.app.outdir, os.path.pardir, "jupyter_execute")
-    )
+    return (Path(env.app.outdir) / os.path.pardir / "jupyter_execute").resolve()
