@@ -4,6 +4,7 @@ import os
 import sys
 from io import StringIO
 from unittest.mock import Mock
+from pathlib import Path
 
 from sphinx.addnodes import download_reference
 from sphinx.testing.util import assert_node, SphinxTestApp, path
@@ -34,16 +35,17 @@ def doctree():
     def doctree(
         source, config=None, return_warnings=False, entrypoint="jupyter_sphinx"
     ):
-        src_dir = tempfile.mkdtemp()
+        src_dir = Path(tempfile.mkdtemp())
         source_trees.append(src_dir)
-        with open(os.path.join(src_dir, "conf.py"), "w", encoding = "utf8") as f:
-            f.write("extensions = ['%s']" % entrypoint)
-            if config is not None:
-                f.write("\n" + config)
-        with open(os.path.join(src_dir, "contents.rst"), "w", encoding = "utf8") as f:
-            f.write(source)
+
+        conf_contents = "extensions = ['%s']" % entrypoint
+        if config is not None:
+            conf_contents += "\n" + config
+        (src_dir / "conf.py").write_text(conf_contents, encoding = "utf8")
+        (src_dir / "contents.rst").write_text(source, encoding = "utf8")
+        
         warnings = StringIO()
-        app = SphinxTestApp(srcdir=path(src_dir), status=StringIO(), warning=warnings)
+        app = SphinxTestApp(srcdir=path(src_dir.as_posix()), status=StringIO(), warning=warnings)
         apps.append(app)
         app.build()
 
@@ -586,6 +588,11 @@ def test_download_role(text, reftarget, caption, tmp_path):
     }
     mock_inliner.configure_mock(**config)
     ret, msg = role('jupyter-download:notebook', text, text, 0, mock_inliner)
+
+    if os.name == "nt":
+        # Get equivalent abs path for Windows
+        reftarget = (Path(tmp_path) / reftarget[1:]).resolve().as_posix()
+
     assert_node(ret[0], [download_reference], reftarget=reftarget)
     assert_node(ret[0][0], [literal, caption])
     assert msg == []
